+++ changelog.md
@@ -7,6 +7,7 @@ Codion Change Log
 - NumberField.numberValue() renamed value().
 - KeyEvents.Builder.enable() and disable() overloaded with Collection parameter.
 - ControlKey.defaultKeyStroke() now Value based and configurable.
+- Controls.Config.create(), ControlMap parameter added, related refactoring.
 ### is.codion.swing.framework.ui
 - EntityTablePanel bug fixed, summary panel toggle control displayed only if a summary panel is available.